Game Card Results
This database information shows the results of the games that various players have played (Rummy, Go Fish, Crazy Eights, Pitch). This database will show the Highest MAX players to the MIN players and the AVERAGE amount that the players have scored throughout the different games and it will tell us who was 1st, 2nd, 3rd or TIE in the different games that were played by players. MAY THE BEST PLAYER WIN.